feat: refactor game logic into GameClient component for improved structure and maintainability
